+++ src/feature/marketplace/ui/listing.document.ts
@@ -8,13 +8,16 @@ export class ListingDocument extends DocumentDto {
   readonly battles: number;
   readonly cardArtUrl: string;
   readonly cardByLevelUrl: string;
+  readonly edition: string;
   readonly fiatSymbol: string;
+  readonly foil: string;
   readonly gold: boolean;
   readonly level: number;
   readonly name: string;
   readonly price: number;
   readonly qty: number;
   readonly rarity: string;
+  readonly role: string;
   readonly splinter: string;
   readonly winpc: number;
 }