+++ CHANGELOG.md
@@ -8,3 +8,7 @@
 
 - ✨ Battle History Page, showing your battle history, or that of any player name
 - ✨ Card Collection Page, show your card collection and its market value
+
+### Fixed
+
+- 🐛 Marketplace filters, easier to use now