+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 <!-- https://keepachangelog.com/en/1.0.0/ -->
 
+## [1.2.2]
+
+### Added
+
+- 🎨 Edition, Foil, Role, Rarity & Splinter now use images for all checkboxes
+
 ## [1.2.1]
 
 ### Added